Fix findings of the new tests: exact animation re-import, atomic imports
- Animation images build runs from color table indices instead of colors,
  so neighbouring pixels with the same color but different indices (team
  colors) survive "import original Stronghold animation" unchanged.
- Gm1Document validates images before changing anything; ReplaceItems
  replaces all items or none.
- Imports read all files first; an unreadable or wrong sized PNG is
  reported with its path and leaves the document unchanged.
- Color table import keeps colors without alpha bit when their cell is
  still transparent, so exporting and importing unchanged tables no longer
  turns them into 0x7FFF.
- Offset patching checks both executables before writing; out of range
  values in Offsets.json are rejected; bounds checks no longer overflow.

+++ Gm1KonverterCrossPlatform.Core/IO/ImageFiles.cs
@@ -25,7 +25,7 @@ namespace Gm1KonverterCrossPlatform.Core.IO
         public static Argb1555Image ToArgb1555(Image<Rgba32> image, int x, int y, int width, int height)
         {
             if (image == null) throw new ArgumentNullException(nameof(image));
-            if (x < 0 || y < 0 || width < 0 || height < 0 || x + width > image.Width || y + height > image.Height)
+            if (x < 0 || y < 0 || width < 0 || height < 0 || (long)x + width > image.Width || (long)y + height > image.Height)
             {
                 throw new ArgumentOutOfRangeException(
                     nameof(image),
